# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Fast Note Sync Service HTTP API.",
        "title": "Fast Note Sync Service HTTP API",
        "contact": {
            "name": "Haierkeys",
            "url": "https://github.com/haierkeys",
            "email": "haierkeys@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/user/change_password": {
            "post": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Handle password change request for current user, validate old password and update new password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Change Password Parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/app.Res"
                        }
                    }
                }
            }
        },
        "/api/user/info": {
            "get": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Handle request to get current user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Res"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "Handle user login HTTP request, validate parameters and return auth token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login Parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Parameters",
                        "schema": {
                            "$ref": "#/definitions/app.Res"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Handle user registration HTTP request, validate parameters and call UserService",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Register Parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Parameters",
                        "schema": {
                            "$ref": "#/definitions/app.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {},
                "data": {},
                "details": {},
                "message": {},
                "status": {
                    "type": "boolean"
                },
                "vault": {}
            }
        },
...
        "dto.UserChangePasswordRequest": {
            "type": "object",
            "required": [
                "confirmPassword",
                "oldPassword",
                "password"
            ],
            "properties": {
                "confirmPassword": {
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UserCreateRequest": {
            "type": "object",
            "required": [
                "confirmPassword",
                "email",
                "password",
                "username"
            ],
            "properties": {
                "confirmPassword": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UserDTO": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "uid": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginRequest": {
            "type": "object",
            "required": [
                "credentials",
                "password"
            ],
            "properties": {
                "credentials": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ShareAuthToken": {
            "type": "apiKey",
            "name": "Share-Token",
            "in": "header"
        },
        "UserAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}